SmartyadsBidAdapter remove usprivacy (#11829)
* SmartyadsBidAdapter remove usprivacy

* refactor

* refactor

* remove unusable code

* try to fix duplicated code

* try to fix duplicated code

* smallfixes

* smallfixes

* smallfix

* tests fix

* revert fixes

* remove language

* move getAdUrlByRegion to libs

* refactor

* undo creative functions

---------
