+++ modules/smartyadsBidAdapter.js
@@ -1,62 +1,14 @@
-import {registerBidder} from '../src/adapters/bidderFactory.js';
+import { registerBidder } from '../src/adapters/bidderFactory.js';
 import { BANNER, NATIVE, VIDEO } from '../src/mediaTypes.js';
 import { config } from '../src/config.js';
 import { convertOrtbRequestToProprietaryNative } from '../src/native.js';
-import { ajax } from '../src/ajax.js';
+import { getAdUrlByRegion } from '../libraries/smartyadsUtils/getAdUrlByRegion.js';
+import { interpretResponse, getUserSyncs } from '../libraries/teqblazeUtils/bidderUtils.js';
 
 const BIDDER_CODE = 'smartyads';
 const GVLID = 534;
-const adUrls = {
-  US_EAST: 'https://n1.smartyads.com/?c=o&m=prebid&secret_key=prebid_js',
-  EU: 'https://n2.smartyads.com/?c=o&m=prebid&secret_key=prebid_js',
-  SGP: 'https://n6.smartyads.com/?c=o&m=prebid&secret_key=prebid_js'
-}
 
 const URL_SYNC = 'https://as.ck-ie.com/prebidjs?p=7c47322e527cf8bdeb7facc1bb03387a';
-
-function isBidResponseValid(bid) {
-  if (!bid.requestId || !bid.cpm || !bid.creativeId ||
-    !bid.ttl || !bid.currency) {
-    return false;
-  }
-  switch (bid['mediaType']) {
-    case BANNER:
-      return Boolean(bid.width && bid.height && bid.ad);
-    case VIDEO:
-      return Boolean(bid.vastUrl) || Boolean(bid.vastXml);
-    case NATIVE:
-      return Boolean(bid.native && bid.native.title && bid.native.image && bid.native.impressionTrackers);
-    default:
-      return false;
-  }
-}
-
-function getAdUrlByRegion(bid) {
-  let adUrl;
-
-  if (bid.params.region && adUrls[bid.params.region]) {
-    adUrl = adUrls[bid.params.region];
-  } else {
-    try {
-      const timezone = Intl.DateTimeFormat().resolvedOptions().timeZone;
-      const region = timezone.split('/')[0];
-
-      switch (region) {
-        case 'Europe':
-          adUrl = adUrls['EU'];
-          break;
-        case 'Asia':
-          adUrl = adUrls['SGP'];
-          break;
-        default: adUrl = adUrls['US_EAST'];
-      }
-    } catch (err) {
-      adUrl = adUrls['US_EAST'];
-    }
-  }
-
-  return adUrl;
-}
 
 export const spec = {
   code: BIDDER_CODE,
@@ -78,8 +30,6 @@ export const spec = {
     let request = {
       'deviceWidth': winTop.screen.width,
       'deviceHeight': winTop.screen.height,
-      'language': (navigator && navigator.language) ? navigator.language : '',
-      'secure': 1,
       'host': location?.domain ?? '',
       'page': location?.page ?? '',
       'coppa': config.getConfig('coppa') === true ? 1 : 0,
@@ -87,11 +37,8 @@ export const spec = {
       'eeid': validBidRequests[0]?.userIdAsEids,
       'ifa': bidderRequest?.ortb2?.device?.ifa,
     };
-    request.language.indexOf('-') != -1 && (request.language = request.language.split('-')[0])
+
     if (bidderRequest) {
-      if (bidderRequest.uspConsent) {
-        request.ccpa = bidderRequest.uspConsent;
-      }
       if (bidderRequest.gdprConsent) {
         request.gdpr = bidderRequest.gdprConsent
       }
@@ -128,59 +75,8 @@ export const spec = {
     }
   },
 
-  interpretResponse: (serverResponse) => {
-    let response = [];
-    serverResponse = serverResponse.body;
-    for (let i = 0; i < serverResponse.length; i++) {
-      let resItem = serverResponse[i];
-      if (isBidResponseValid(resItem)) {
-        response.push(resItem);
-      }
-    }
-    return response;
-  },
-
-  getUserSyncs: (syncOptions, serverResponses = [], gdprConsent = {}, uspConsent = '', gppConsent = '') => {
-    let syncs = [];
-    let { gdprApplies, consentString = '' } = gdprConsent;
-
-    if (syncOptions.iframeEnabled) {
-      syncs.push({
-        type: 'iframe',
-        url: `${URL_SYNC}&gdpr=${gdprApplies ? 1 : 0}&gdpr_consent=${consentString}&type=iframe&us_privacy=${uspConsent}&gpp=${gppConsent}`
-      });
-    } else {
-      syncs.push({
-        type: 'image',
-        url: `${URL_SYNC}&gdpr=${gdprApplies ? 1 : 0}&gdpr_consent=${consentString}&type=image&us_privacy=${uspConsent}&gpp=${gppConsent}`
-      });
-    }
-
-    return syncs
-  },
-
-  onBidWon: function(bid) {
-    if (bid.winUrl) {
-      ajax(bid.winUrl, () => {}, JSON.stringify(bid));
-    } else {
-      if (bid?.postData && bid?.postData[0] && bid?.postData[0].params && bid?.postData[0].params[0].host == 'prebid') {
-        ajax('https://et-nd43.itdsmr.com/?c=o&m=prebid&secret_key=prebid_js&winTest=1', () => {}, JSON.stringify(bid));
-      }
-    }
-  },
-
-  onTimeout: function(bid) {
-    if (bid?.postData && bid?.postData[0] && bid?.postData[0].params && bid?.postData[0].params[0].host == 'prebid') {
-      ajax('https://et-nd43.itdsmr.com/?c=o&m=prebid&secret_key=prebid_js&bidTimeout=1', () => {}, JSON.stringify(bid));
-    }
-  },
-
-  onBidderError: function(bid) {
-    if (bid?.postData && bid?.postData[0] && bid?.postData[0].params && bid?.postData[0].params[0].host == 'prebid') {
-      ajax('https://et-nd43.itdsmr.com/?c=o&m=prebid&secret_key=prebid_js&bidderError=1', () => {}, JSON.stringify(bid));
-    }
-  },
-
+  interpretResponse,
+  getUserSyncs: getUserSyncs(URL_SYNC),
 };
 
 registerBidder(spec);
