+++ test/spec/modules/smartyadsBidAdapter_spec.js
@@ -61,12 +61,10 @@ describe('SmartyadsAdapter', function () {
     it('Returns valid data if array of bids is valid', function () {
       let data = serverRequest.data;
       expect(data).to.be.an('object');
-      expect(data).to.have.all.keys('deviceWidth', 'deviceHeight', 'language', 'secure', 'host', 'page', 'placements', 'coppa', 'eeid', 'ifa');
+      expect(data).to.have.all.keys('deviceWidth', 'deviceHeight', 'host', 'page', 'placements', 'coppa', 'eeid', 'ifa');
       expect(data.deviceWidth).to.be.a('number');
       expect(data.deviceHeight).to.be.a('number');
       expect(data.coppa).to.be.a('number');
-      expect(data.language).to.be.a('string');
-      expect(data.secure).to.be.within(0, 1);
       expect(data.host).to.be.a('string');
       expect(data.page).to.be.a('string');
       let placement = data['placements'][0];
@@ -126,8 +124,8 @@ describe('SmartyadsAdapter', function () {
       expect(dataItem.width).to.equal(300);
       expect(dataItem.height).to.equal(250);
       expect(dataItem.ad).to.equal('Test');
-      expect(dataItem.meta).to.have.property('advertiserDomains')
-      expect(dataItem.meta.advertiserDomains).to.deep.equal(['example.com']);
+      expect(dataItem.meta).to.have.property('advertiserDomains');
+      expect(dataItem.meta.advertiserDomains).to.be.an('array');
       expect(dataItem.ttl).to.equal(120);
       expect(dataItem.creativeId).to.equal('2');
       expect(dataItem.netRevenue).to.be.true;
@@ -152,7 +150,7 @@ describe('SmartyadsAdapter', function () {
 
       let dataItem = videoResponses[0];
       expect(dataItem).to.have.all.keys('requestId', 'cpm', 'vastUrl', 'ttl', 'creativeId',
-        'netRevenue', 'currency', 'dealId', 'mediaType');
+        'netRevenue', 'currency', 'dealId', 'mediaType', 'meta');
       expect(dataItem.requestId).to.equal('23fhj33i987f');
       expect(dataItem.cpm).to.equal(0.5);
       expect(dataItem.vastUrl).to.equal('test.com');
@@ -183,7 +181,7 @@ describe('SmartyadsAdapter', function () {
       expect(nativeResponses).to.be.an('array').that.is.not.empty;
 
       let dataItem = nativeResponses[0];
-      expect(dataItem).to.have.keys('requestId', 'cpm', 'ttl', 'creativeId', 'netRevenue', 'currency', 'mediaType', 'native');
+      expect(dataItem).to.have.keys('requestId', 'cpm', 'ttl', 'creativeId', 'netRevenue', 'currency', 'mediaType', 'native', 'meta');
       expect(dataItem.native).to.have.keys('clickUrl', 'impressionTrackers', 'title', 'image')
       expect(dataItem.requestId).to.equal('23fhj33i987f');
       expect(dataItem.cpm).to.equal(0.4);
@@ -263,7 +261,7 @@ describe('SmartyadsAdapter', function () {
     });
   });
   describe('getUserSyncs', function () {
-    const syncUrl = 'https://as.ck-ie.com/prebidjs?p=7c47322e527cf8bdeb7facc1bb03387a&gdpr=0&gdpr_consent=&type=iframe&us_privacy=&gpp=';
+    const syncUrl = 'https://as.ck-ie.com/prebidjs?p=7c47322e527cf8bdeb7facc1bb03387a/iframe?pbjs=1&coppa=0';
     const syncOptions = {
       iframeEnabled: true
     };
@@ -275,81 +273,6 @@ describe('SmartyadsAdapter', function () {
       expect(userSync).to.deep.equal([
         { type: 'iframe', url: syncUrl }
       ]);
-    });
-  });
-
-  describe('onBidWon', function () {
-    it('should exists', function () {
-      expect(spec.onBidWon).to.exist.and.to.be.a('function');
-    });
-
-    it('should send a valid bid won notice', function () {
-      const bid = {
-        'c': 'o',
-        'm': 'prebid',
-        'secret_key': 'prebid_js',
-        'winTest': '1',
-        'postData': [{
-          'bidder': 'smartyads',
-          'params': [
-            {'host': 'prebid',
-              'accountid': '123',
-              'sourceid': '12345'
-            }]
-        }]
-      };
-      spec.onBidWon(bid);
-      expect(server.requests.length).to.equal(1);
-    });
-  });
-
-  describe('onTimeout', function () {
-    it('should exists', function () {
-      expect(spec.onTimeout).to.exist.and.to.be.a('function');
-    });
-
-    it('should send a valid bid timeout notice', function () {
-      const bid = {
-        'c': 'o',
-        'm': 'prebid',
-        'secret_key': 'prebid_js',
-        'bidTimeout': '1',
-        'postData': [{
-          'bidder': 'smartyads',
-          'params': [
-            {'host': 'prebid',
-              'accountid': '123',
-              'sourceid': '12345'
-            }]
-        }]
-      };
-      spec.onTimeout(bid);
-      expect(server.requests.length).to.equal(1);
-    });
-  });
-
-  describe('onBidderError', function () {
-    it('should exists', function () {
-      expect(spec.onBidderError).to.exist.and.to.be.a('function');
-    });
-
-    it('should send a valid bidder error notice', function () {
-      const bid = {
-        'c': 'o',
-        'm': 'prebid',
-        'secret_key': 'prebid_js',
-        'bidderError': '1',
-        'postData': [{
-          'bidder': 'smartyads',
-          'params': [
-            {'host': 'prebid',
-              'accountid': '123',
-              'sourceid': '12345'
-            }]
-        }]
-      };
-      spec.onBidderError(bid);
-      expect(server.requests.length).to.equal(1);
     });
   });
 });
